+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.10.0]
 ### Changed
 - [issues/29](https://github.com/nasa/python_cmr/issues/29) - Date parsing has been improved to accept more ISO-8601 string formats as well as timezone-aware datetime objects
 ### Added
@@ -52,7 +54,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Older]
 - Prior releases of this software originated from https://github.com/jddeal/python-cmr/releases
 
-[Unreleased]: https://github.com/nasa/python_cmr/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/nasa/python_cmr/compare/v0.10.0...HEAD
+[0.10.0]: https://github.com/nasa/python_cmr/compare/v0.9.0...v0.10.0
+[0.9.0]: https://github.com/nasa/python_cmr/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/nasa/python_cmr/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/nasa/python_cmr/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/nasa/python_cmr/compare/v0.5.0...v0.6.0